Added interfaces. Support for fixed and resizable display modes work perfectly.

+++ rt5e/core/src/org/lazaro/rt5e/logic/utility/NodeCollection.java
@@ -76,7 +76,7 @@ public class NodeCollection<N extends Node> extends AbstractCollection<N> {
             return false;
         }
         nodes[index] = node;
-        indexes.add((Integer) index);
+        indexes.add(index);
         node.setIndex(index);
         return true;
     }
